A3C example

+++ Holodeck/HolodeckEnvironment.py
@@ -7,7 +7,7 @@ import os
 
 class HolodeckEnvironment(object):
 
-    def __init__(self, agent_type, agent_name, task_key=None, height=256, width=256, verbose=False, grayscale=False):
+    def __init__(self, agent_type, agent_name, task_key=None, height=256, width=256, verbose=False, grayscale=False,hostname="localhost",startWorld=True):
         self.resolution = (height, width, 1 if grayscale else 3)
         self.grayscale = grayscale
         self.verbose = verbose
@@ -25,9 +25,11 @@ class HolodeckEnvironment(object):
         if self.verbose:
             print "starting process"
 
-        self.world_process = subprocess.Popen([task_map[task_key], '-opengl4', '-SILENT', '-LOG=MyLog.txt', '-ResX=' + str(width), " -ResY=" + str(height)],
-                                               stdout=open(os.devnull, 'w') if not verbose else None,
-                                               stderr=open(os.devnull, 'w') if not verbose else None)
+        if startWorld:
+            self.world_process = subprocess.Popen([task_map[task_key], '-opengl4', '-SILENT', '-LOG=MyLog.txt', '-ResX=' + str(width), " -ResY=" + str(height)],
+                                                   stdout=open(os.devnull, 'w') if not verbose else None,
+                                                   stderr=open(os.devnull, 'w') if not verbose else None)
+
         if self.verbose:
             print "process started"
 
@@ -36,11 +38,12 @@ class HolodeckEnvironment(object):
         if self.verbose:
             print "process registered for exit"
         
-        self.agent = agent_type(hostname='localhost', port=8989, agentName=agent_name, height=height, width=width,grayscale=grayscale)
+        self.agent = agent_type(hostname=hostname, port=8989, agentName=agent_name, height=height, width=width,grayscale=grayscale)
         self.agent.wait_for_connect()
 
     def __on_exit__(self):
-        self.world_process.kill()
+        if hasattr(self, 'world_process'):
+            self.world_process.kill()
 
     @property
     def action_space(self):
@@ -83,28 +86,28 @@ class HolodeckUAVEnvironment(HolodeckEnvironment):
         return spaces.Box(0, 255, shape=self.resolution)
 
 class HolodeckUAVMazeWorld(HolodeckEnvironment):
-    def __init__(self, agent_name="UAV0", verbose=False, resolution=(32, 32),grayscale=False):
+    def __init__(self, agent_name="UAV0", verbose=False, resolution=(32, 32),grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckUAVMazeWorld, self).__init__(agent_name=agent_name, verbose=verbose, task_key="MazeWorld_UAV",
                                                      height=resolution[0], width=resolution[1],
-                                                     agent_type=Holodeck.SimulatorAgent.UAVAgent,grayscale=grayscale)
-        self.state_sensors = ['PrimaryPlayerCamera']
-
-    @property
-    def observation_space(self):
-        return spaces.Box(0, 255, shape=self.resolution)
-
-class HolodeckUAVForestWorld(HolodeckEnvironment):
-    def __init__(self, agent_name="UAV0", verbose=False, resolution=(32, 32),grayscale=False):
-        super(HolodeckUAVForestWorld, self).__init__(agent_name=agent_name, verbose=verbose, task_key="ForestWorld_UAV",
-                                                     height=resolution[0], width=resolution[1],
-                                                     agent_type=Holodeck.SimulatorAgent.UAVAgent,grayscale=grayscale)
+                                                     agent_type=Holodeck.SimulatorAgent.UAVAgent,grayscale=grayscale,
+                                                     hostname=hostname,startWorld=startWorld)
         self.state_sensors = ['PrimaryPlayerCamera',"OrientationSensor"]
 
     @property
     def observation_space(self):
         return spaces.Box(0, 255, shape=self.resolution)
 
+class HolodeckUAVForestWorld(HolodeckEnvironment):
+    def __init__(self, agent_name="UAV0", verbose=False, resolution=(32, 32),grayscale=False,hostname="localhost",startWorld=True):
+        super(HolodeckUAVForestWorld, self).__init__(agent_name=agent_name, verbose=verbose, task_key="ForestWorld_UAV",
+                                                     height=resolution[0], width=resolution[1],
+                                                     agent_type=Holodeck.SimulatorAgent.UAVAgent,grayscale=grayscale,
+                                                     hostname=hostname,startWorld=startWorld)
+        self.state_sensors = ['PrimaryPlayerCamera',"OrientationSensor"]
 
+    @property
+    def observation_space(self):
+        return spaces.Box(0, 255, shape=self.resolution)
 
 class HolodeckContinuousSphereEnvironment(HolodeckEnvironment):
     def __init__(self, agent_name="sphere0", verbose=False,grayscale=False):
@@ -117,10 +120,11 @@ class HolodeckContinuousSphereEnvironment(HolodeckEnvironment):
         return spaces.Box(0, 255, shape=self.resolution)
 
 class HolodeckContinuousSphereMazeWorld(HolodeckEnvironment):
-    def __init__(self, agent_name="sphere0", verbose=False, resolution=(32,32),grayscale=False):
+    def __init__(self, agent_name="sphere0", verbose=False, resolution=(32,32),grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckContinuousSphereMazeWorld, self).__init__(agent_name=agent_name, verbose=verbose,task_key="MazeWorld_sphere",
                                                                     height=resolution[0],width=resolution[1],
-                                                                  agent_type=Holodeck.SimulatorAgent.ContinuousSphereAgent,grayscale=grayscale)
+                                                                  agent_type=Holodeck.SimulatorAgent.ContinuousSphereAgent,grayscale=grayscale,
+                                                                  hostname=hostname,startWorld=startWorld)
         self.state_sensors = ['PrimaryPlayerCamera']
 
     @property
@@ -131,9 +135,10 @@ class HolodeckContinuousSphereMazeWorld(HolodeckEnvironment):
 
 
 class HolodeckDiscreteSphereEnvironment(HolodeckEnvironment):
-    def __init__(self, agent_name="sphere0", verbose=False,grayscale=False):
+    def __init__(self, agent_name="sphere0", verbose=False,grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckDiscreteSphereEnvironment, self).__init__(agent_name=agent_name, verbose=verbose,
-                                                                agent_type=Holodeck.SimulatorAgent.DiscreteSphereAgent,grayscale=grayscale)
+                                                                agent_type=Holodeck.SimulatorAgent.DiscreteSphereAgent,grayscale=grayscale,
+                                                                hostname=hostname,startWorld=startWorld)
         self.state_sensors = ['PrimaryPlayerCamera']
 
     @property
@@ -141,10 +146,11 @@ class HolodeckDiscreteSphereEnvironment(HolodeckEnvironment):
         return spaces.Box(0, 255, shape=self.resolution)
 
 class HolodeckDiscreteSphereMazeWorld(HolodeckEnvironment):
-    def __init__(self, agent_name="sphere0", verbose=False,resolution=(32,32),grayscale=False):
+    def __init__(self, agent_name="sphere0", verbose=False,resolution=(32,32),grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckDiscreteSphereMazeWorld, self).__init__(agent_name=agent_name, verbose=verbose,task_key="MazeWorld_sphere",
                                                                     height=resolution[0],width=resolution[1],
-                                                                agent_type=Holodeck.SimulatorAgent.DiscreteSphereAgent,grayscale=grayscale)
+                                                                agent_type=Holodeck.SimulatorAgent.DiscreteSphereAgent,grayscale=grayscale,
+                                                                hostname=hostname,startWorld=startWorld)
         self.state_sensors = ['PrimaryPlayerCamera']
 
     @property
@@ -153,9 +159,10 @@ class HolodeckDiscreteSphereMazeWorld(HolodeckEnvironment):
 
 
 class HolodeckAndroidEnvironment(HolodeckEnvironment):
-    def __init__(self, agent_name="android0", verbose=False,grayscale=False):
+    def __init__(self, agent_name="android0", verbose=False,grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckAndroidEnvironment, self).__init__(agent_name=agent_name, verbose=verbose,
-                                                         agent_type=Holodeck.SimulatorAgent.AndroidAgent,grayscale=grayscale)
+                                                         agent_type=Holodeck.SimulatorAgent.AndroidAgent,grayscale=grayscale,
+                                                         hostname=hostname,startWorld=startWorld)
 
         # self.agent.send_command('AndroidConfiguration', {"AreCollisionsVisible": True})
         # self.state_sensors = ['PrimaryPlayerCamera', 'IMUSensor', 'JointRotationSensor', 'RelativeSkeletalPositionSensor']
@@ -166,10 +173,11 @@ class HolodeckAndroidEnvironment(HolodeckEnvironment):
         return spaces.Box(0, 255, shape=self.resolution)
 
 class HolodeckAndroidExampleWorldEnvironment(HolodeckEnvironment):
-    def __init__(self, agent_name="android0", verbose=False,resolution=(32,32),grayscale=False):
+    def __init__(self, agent_name="android0", verbose=False,resolution=(32,32),grayscale=False,hostname="localhost",startWorld=True):
         super(HolodeckAndroidExampleWorldEnvironment, self).__init__(agent_name=agent_name, verbose=verbose,task_key="ExampleWorld_android",
                                                                     height=resolution[0],width=resolution[1],
-                                                                    agent_type=Holodeck.SimulatorAgent.AndroidAgent,grayscale=grayscale)
+                                                                    agent_type=Holodeck.SimulatorAgent.AndroidAgent,grayscale=grayscale,
+                                                                    hostname=hostname,startWorld=startWorld)
 
         # self.agent.send_command('AndroidConfiguration', {"AreCollisionsVisible": True})
         # self.state_sensors = ['PrimaryPlayerCamera', 'IMUSensor', 'JointRotationSensor', 'RelativeSkeletalPositionSensor']
